+++ packages/experimental/Text/CHANGELOG.md
@@ -1,8 +1,48 @@
 # Change Log - @fluentui-react-native/experimental-text
 
-This log was last generated on Fri, 06 Aug 2021 00:36:02 GMT and should not be manually modified.
+This log was last generated on Mon, 09 Aug 2021 22:49:01 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.6.7
+
+Mon, 09 Aug 2021 22:49:01 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/framework to v0.5.26 (ruaraki@microsoft.com)
+
+## 0.6.6
+
+Mon, 09 Aug 2021 19:16:08 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/framework to v0.5.25 (ruaraki@microsoft.com)
+
+## 0.6.5
+
+Sat, 07 Aug 2021 02:50:29 GMT
+
+### Patches
+
+- Merge branch 'master' of https://github.com/microsoft/fluentui-react-native into win32Shimmer (patboyd@microsoft.com)
+
+## 0.6.4
+
+Sat, 07 Aug 2021 00:40:04 GMT
+
+### Patches
+
+- enable usePressableState with stock Pressable component (jasonmo@microsoft.com)
+
+## 0.6.3
+
+Fri, 06 Aug 2021 23:18:21 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/framework to v0.5.23 (ruaraki@microsoft.com)
 
 ## 0.6.2